+++ packages/terra-form-select/tests/wdio/select-field-spec.js
@@ -52,7 +52,7 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     Terra.it.validatesElement();
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('#select-field');
+      browser.click('[class*="tag"]');
     });
 
     Terra.it.validatesElement('open-dropdown', { selector: '#root' });
@@ -105,7 +105,7 @@ Terra.describeViewports('Select Field', ['tiny'], () => {
     Terra.it.validatesElement();
 
     it('should open the dropdown by clicking the toggle', () => {
-      browser.click('#select-field');
+      browser.click('[class*="tag"]');
     });
 
     Terra.it.matchesScreenshot('open-dropdown', { selector: '#root' });